+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/fivetran/go-fivetran/compare/v1.0.6...HEAD)
+## [Unreleased](https://github.com/fivetran/go-fivetran/compare/v1.0.7...HEAD)
+
+## [1.0.7](https://github.com/fivetran/go-fivetran/compare/v1.0.6...v1.0.7)
+
+Supported the following Fivetran API endpoints:
+- [List all destinations within account](https://fivetran.com/docs/rest-api/api-reference/destinations/list-destinations)
+- [List all connectors within account](https://fivetran.com/docs/rest-api/api-reference/connectors/list-connectors)
+- [List all Log Services](https://fivetran.com/docs/rest-api/api-reference/log-service-management/list-log-services)
 
 ## [1.0.6](https://github.com/fivetran/go-fivetran/compare/v1.0.5...v1.0.6)